refactor: Update repository implementations to use record accessors
- Replace getter methods with record accessor methods in JDBC repositories
- Update PetJdbcRepository and PetLookUpService to use pet.identity().id() instead of pet.getIdentity().getId()
- Fix import statements to reference reorganized infrastructure packages
